Fix implicit conversion

+++ PAC/common/imitation_TE.cpp
@@ -13,7 +13,7 @@ float imitation_TE::get_random()
     return  fv.at(index_fv) *0.1f + iv.at(index_iv);
     }
 
-void imitation_TE::set_max(int max_in_range)
+void imitation_TE::set_max(float max_in_range)
     {
     this->max = max_in_range;
     }
@@ -23,7 +23,7 @@ float imitation_TE::get_max() const
     return max;
     }
 
-void imitation_TE::set_min(int min_in_range)
+void imitation_TE::set_min(float min_in_range)
     {
     this->min = min_in_range;
     }
